- NSInputStream fixes for NSData based stream - NSBitmapImageRep initWithData crash fixed due to uninitialized _bitmapPlanes - Changed all CGImageSource's to use a data provider instead of data

+++ Foundation/NSStream/NSInputStream_data.h
@@ -13,7 +13,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
    NSError       *_error;
    NSStreamStatus _status;
    NSData        *_data;
-   unsigned       _position;
+   long long      _position;
 }
 
 @end
